Added the Python programs from before repo
All of the Python programs, including the bats to run them, are being
commited by this here commit, in order to allow for them to be part of
the repo. All TXT files are not required, and are only used to store
info used for the programs that have one.